backend/requirements2.txt
aiohttp==3.8.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bitarray==2.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
cssselect2==0.7.0
cytoolz==0.12.1
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==4.0.2
#django-4-jet==1.0.9
django-jet-reboot
django-allauth==0.52.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
environ==1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
hexbytes==0.3.0
html5lib==1.1
idna==3.4
ipfshttpclient==0.8.0a2
jsonfield==3.1.0
jsonschema==4.17.3
lru-dict==1.1.8
lxml==4.9.2
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
oauthlib==3.2.2
oscrypto==1.3.0
parsimonious==0.8.1
phonenumberslite==8.13.6
Pillow==9.0.1
protobuf==3.19.5
pycparser==2.21
pycryptodome==3.17
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.6.0
PyPDF3==1.0.6
pyrsistent==0.19.3
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
#pywin32==305
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
rlp==2.0.1
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
toolz==0.12.0
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.14
varint==1.0.2
web3==5.31.3
webencodings==0.5.1
websockets==9.1
xhtml2pdf==0.2.8
yarl==1.8.2
backports.zoneinfo
psycopg2-binary
gunicorn
